+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     Previously: `asfa push --alias my-alias.txt -- my-original-file.txt`
     Now: `asfa push --alias my-alias.txt my-original-file.txt`
   * Check response for errors regaring `atd` not running if `--expire` specified.
+  * Now supports a default setting for `--expire` in the config. Same as auth,
+    both a global setting and host-specific settings are supported.
 * `list`:
   * More informative message if no files are present on remote site.
 * tests: